refactor(bundler): Move `GemInfo` to its own file for a better overview
